Epic/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-nested-routers==0.93.4
psycopg2==2.9.5
PyJWT==2.4.0
pytz==2022.6
sqlparse==0.4.3
